test/e2e: add deploying separate OAuth server to util
This allows to create an OAuth server pod into the namespace of the
running e2e tests and a bunch of util functions to simpify getting tokens
from such a server.

Note that the OpenShift integrated OAuth server tends to create
or modify cluster-scoped OAuthClient objects so some hacking
needed to be done in order to revert these back to their original
form and also in order to still make it possible to deploy
multiple OAuth servers in parallel.

Co-authored-by: Sally O'Malley <somalley@redhat.com>